+++ packages/vulcan-lib/lib/server/apollo-ssr/components/AppGenerator.jsx
@@ -1,3 +1,6 @@
+/**
+ * The App + relevant wrappers
+ */
 import React from 'react';
 import { ApolloProvider } from 'react-apollo';
 import { StaticRouter } from 'react-router';
@@ -7,11 +10,10 @@ import { StaticRouter } from 'react-router';
 // so Components.App is not defined here
 import { Components } from 'meteor/vulcan:lib'
 
-// TODO: adapt to Vulcan
 // The client-side App will instead use <BrowserRouter>
 // see client-side vulcan:core/lib/client/start.jsx implementation
 // we do the same server side
-const appGenerator = ({ req, client, context }) => {
+const AppGenerator = ({ req, client, context }) => {
   const App = (
     <ApolloProvider client={client}>
       <StaticRouter location={req.url} context={context}>
@@ -21,4 +23,4 @@ const appGenerator = ({ req, client, context }) => {
   );
   return App;
 };
-export default appGenerator;
+export default AppGenerator;
