--- packages/vulcan-lib/lib/server/apollo-ssr/start.js
@@ -1,12 +1,0 @@
-import {
-  populateComponentsApp,
-  populateRoutesApp,
-  initializeFragments
-} from 'meteor/vulcan:lib';
-
-Meteor.startup(() => {
-  // init the application components and routes, including components & routes from 3rd-party packages
-  initializeFragments();
-  populateComponentsApp();
-  populateRoutesApp();
-});
